Day20-25-production-rag/requirements.txt
openai
oracledb
python-dotenv
numpy
fastapi
uvicorn
pydantic
pydantic-settings
rank-bm25
python-multipart
pandas
drain3

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0